+++ units/units_math.hpp
@@ -15,14 +15,16 @@ SPDX-License-Identifier: BSD-3-Clause
 
 namespace UNITS_NAMESPACE {
 
-    /** declare the type traits*/
+/** declare the type traits*/
 
 /// type_trait for detecting any measurement type
 template<typename T>
 struct is_measurement : std::false_type {
 };
 
-template<> struct is_measurement<measurement> : std::true_type {};
+template<>
+struct is_measurement<measurement> : std::true_type {
+};
 template<>
 struct is_measurement<precise_measurement> : std::true_type {
 };
@@ -66,81 +68,95 @@ struct is_unit<precise_unit> : std::true_type {
 
 template<
     typename X,
-    typename = typename std::enable_if<is_measurement<X>::value>::type> X cbrt(const X& measure)
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
+X cbrt(const X& measure)
 {
-    return root(measure,3);
+    return root(measure, 3);
 }
 
-template<typename X, typename= typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 X floor(const X& measure)
 {
-    return X((std::floor)(measure.value()),measure.units());
+    return X((std::floor)(measure.value()), measure.units());
 }
 
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 X ceil(const X& measure)
 {
-    return X((std::ceil)(measure.value()),measure.units());
+    return X((std::ceil)(measure.value()), measure.units());
 }
 
-
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 X round(const X& measure)
 {
-    return X((std::round)(measure.value()),measure.units());
+    return X((std::round)(measure.value()), measure.units());
 }
 
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 X trunc(const X& measure)
 {
-    return X((std::trunc)(measure.value()),measure.units());
+    return X((std::trunc)(measure.value()), measure.units());
 }
-
 
 template<
     typename X,
     typename Y,
-    typename = typename std::enable_if<is_measurement<X>::value || is_measurement<Y>::value >::type>
+    typename = typename std::enable_if<
+        is_measurement<X>::value || is_measurement<Y>::value>::type>
 auto hypot(const X& measure1, const Y& measure2)
 {
-    return sqrt(pow(measure1,2)+pow(measure2,2));
+    return sqrt(pow(measure1, 2) + pow(measure2, 2));
 }
 
 template<
     typename X,
     typename Y,
     typename Z,
-    typename = typename std::enable_if<is_measurement<X>::value ||
-        is_measurement<Y>::value || is_measurement<Z>::value>::type>
+    typename = typename std::enable_if<
+        is_measurement<X>::value || is_measurement<Y>::value ||
+        is_measurement<Z>::value>::type>
 auto hypot(const X& measure1, const Y& measure2, const Z& measure3)
 {
-    return sqrt(pow(measure1,2) + pow(measure2,2)+pow(measure3,2));
+    return sqrt(pow(measure1, 2) + pow(measure2, 2) + pow(measure3, 2));
 }
 
 template<
     typename X,
     typename Y,
-    typename =
-        typename std::enable_if<is_measurement<X>::value || is_measurement<Y>::value>::type>
+    typename = typename std::enable_if<
+        is_measurement<X>::value || is_measurement<Y>::value>::type>
 auto fmod(const X& measure1, const Y& measure2)
 {
-    return measure1-trunc(measure1/measure2)*measure2;
+    return measure1 - trunc(measure1 / measure2) * measure2;
 }
 
-
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 double sin(const X& measure)
 {
     return (std::sin)(measure.value_as(precise::rad));
 }
 
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 double cos(const X& measure)
 {
     return (std::cos)(measure.value_as(precise::rad));
 }
 
-template<typename X, typename = typename std::enable_if<is_measurement<X>::value>::type>
+template<
+    typename X,
+    typename = typename std::enable_if<is_measurement<X>::value>::type>
 double tan(const X& measure)
 {
     return (std::tan)(measure.value_as(precise::rad));